--- a/decoder/assembler_amd64_go116.go
+++ b/decoder/assembler_amd64_go116.go
@@ -70,7 +70,7 @@ const (
     _FP_args   = 96     // 96 bytes to pass arguments and return values for this function
     _FP_fargs  = 80     // 80 bytes for passing arguments to other Go functions
     _FP_saves  = 40     // 40 bytes for saving the registers before CALL instructions
-    _FP_locals = 120    // 120 bytes for local variables
+    _FP_locals = 144    // 144 bytes for local variables
 )
 
 const (
@@ -101,6 +101,7 @@ const (
     _LB_unquote_error   = "_unquote_error"
     _LB_parsing_error   = "_parsing_error"
     _LB_parsing_error_v = "_parsing_error_v"
+    _LB_mismatch_error   = "_mismatch_error"
 )
 
 const (
@@ -112,6 +113,8 @@ const (
     _LB_char_m2_error = "_char_m2_error"
     _LB_char_m3_error = "_char_m3_error"
 )
+
+const _LB_skip_one = "_skip_one"
 
 var (
     _AX = jit.Reg("AX")
@@ -196,6 +199,12 @@ var (
 
 var _VAR_fl = jit.Ptr(_SP, _FP_fargs + _FP_saves + 112)
 
+var (
+    _VAR_et = jit.Ptr(_SP, _FP_fargs + _FP_saves + 120) // save dismatched type
+    _VAR_ic = jit.Ptr(_SP, _FP_fargs + _FP_saves + 128) // save dismatched position
+    _VAR_pc = jit.Ptr(_SP, _FP_fargs + _FP_saves + 136) // save skip return pc
+)
+
 type _Assembler struct {
     jit.BaseAssembler
     p _Program
@@ -224,6 +233,8 @@ func (self *_Assembler) compile() {
     self.copy_string()
     self.escape_string()
     self.escape_string_twice()
+    self.skip_one()
+    self.mismatch_error()
     self.type_error()
     self.field_error()
     self.range_error()
@@ -295,6 +306,12 @@ var _OpFuncTab = [256]func(*_Assembler, *_Instr) {
     _OP_recurse          : (*_Assembler)._asm_OP_recurse,
     _OP_goto             : (*_Assembler)._asm_OP_goto,
     _OP_switch           : (*_Assembler)._asm_OP_switch,
+    _OP_check_bool       : (*_Assembler)._asm_OP_check_bool,
+    _OP_check_num        : (*_Assembler)._asm_OP_check_num,
+    _OP_check_char_0     : (*_Assembler)._asm_OP_check_char_0,
+    _OP_dismatch_err     : (*_Assembler)._asm_OP_dismatch_err,
+    _OP_go_skip          : (*_Assembler)._asm_OP_go_skip,
+    _OP_add              : (*_Assembler)._asm_OP_add,
 }
 
 func (self *_Assembler) instr(v *_Instr) {
@@ -315,8 +332,10 @@ func (self *_Assembler) instrs() {
 
 func (self *_Assembler) epilogue() {
     self.Mark(len(self.p))
-    self.Emit("XORL", _ET, _ET)                     // XORL ET, ET
     self.Emit("XORL", _EP, _EP)                     // XORL EP, EP
+    self.Emit("MOVQ", _VAR_et, _ET)                 // MOVQ VAR_et, ET
+    self.Emit("TESTQ", _ET, _ET)                    // TESTQ ET, ET
+    self.Sjmp("JNZ", _LB_mismatch_error)            // JNZ _LB_mismatch_error
     self.Link(_LB_error)                            // _error:
     self.Emit("MOVQ", _IC, _RET_rc)                 // MOVQ IC, rc<>+40(FP)
     self.Emit("MOVQ", _ET, _RET_et)                 // MOVQ ET, et<>+48(FP)
@@ -339,6 +358,7 @@ func (self *_Assembler) prologue() {
     self.Emit("MOVQ", jit.Imm(_MaxDigitNums), _VAR_st_Dc)    // MOVQ $_MaxDigitNums, ss.Dcap
     self.Emit("LEAQ", jit.Ptr(_ST, _DbufOffset), _AX)           // LEAQ _DbufOffset(ST), AX
     self.Emit("MOVQ", _AX, _VAR_st_Db)                          // MOVQ AX, ss.Dbuf
+    self.Emit("MOVQ", jit.Imm(0), _VAR_et)                          // MOVQ AX, ss.Dp
 }
 
 /** Function Calling Helpers **/
@@ -404,11 +424,12 @@ func (self *_Assembler) call_vf(fn obj.Addr) {
 /** Assembler Error Handlers **/
 
 var (
-    _F_convT64     = jit.Func(convT64)
-    _F_error_wrap  = jit.Func(error_wrap)
-    _F_error_type  = jit.Func(error_type)
-    _F_error_field = jit.Func(error_field)
-    _F_error_value = jit.Func(error_value)
+    _F_convT64        = jit.Func(convT64)
+    _F_error_wrap     = jit.Func(error_wrap)
+    _F_error_type     = jit.Func(error_type)
+    _F_error_field    = jit.Func(error_field)
+    _F_error_value    = jit.Func(error_value)
+    _F_error_mismatch = jit.Func(error_mismatch)
 )
 
 var (
@@ -438,6 +459,46 @@ func (self *_Assembler) type_error() {
     self.Emit("MOVQ", jit.Ptr(_SP, 8), _ET)     // MOVQ    8(SP), ET
     self.Emit("MOVQ", jit.Ptr(_SP, 16), _EP)    // MOVQ    16(SP), EP
     self.Sjmp("JMP" , _LB_error)                // JMP     _error
+}
+
+
+func (self *_Assembler) mismatch_error() {
+    self.Link(_LB_mismatch_error)               // _type_error:
+    self.Emit("MOVQ", _ARG_sp, _AX)
+    self.Emit("MOVQ", _AX, jit.Ptr(_SP, 0))     // MOVQ    AX, (SP)
+    self.Emit("MOVQ", _ARG_sl, _CX)
+    self.Emit("MOVQ", _CX, jit.Ptr(_SP, 8))     // MOVQ    CX, 8(SP)
+    self.Emit("MOVQ", _VAR_ic, _AX)
+    self.Emit("MOVQ", _AX, jit.Ptr(_SP, 16))    // MOVQ    AX, 16(SP)
+    self.Emit("MOVQ", _VAR_et, _CX)
+    self.Emit("MOVQ", _CX, jit.Ptr(_SP, 24))    // MOVQ    CX, 24(SP)
+    self.call_go(_F_error_mismatch)             // CALL_GO error_type
+    self.Emit("MOVQ", jit.Ptr(_SP, 32), _ET)    // MOVQ    32(SP), ET
+    self.Emit("MOVQ", jit.Ptr(_SP, 40), _EP)    // MOVQ    40(SP), EP
+    self.Sjmp("JMP" , _LB_error)                // JMP     _error
+}
+
+func (self *_Assembler) _asm_OP_dismatch_err(p *_Instr) {
+    self.Emit("MOVQ", _IC, _VAR_ic)           
+    self.Emit("MOVQ", jit.Type(p.vt()), _AX)  
+    self.Emit("MOVQ", _AX, _VAR_et)
+}
+
+func (self *_Assembler) _asm_OP_go_skip(p *_Instr) {
+    self.Byte(0x4c, 0x8d, 0x0d)         // LEAQ (PC), R9
+    self.Xref(p.vi(), 4)
+    self.Emit("MOVQ", _R9, _VAR_pc)
+    self.Sjmp("JMP"  , _LB_skip_one)            // JMP     _skip_one
+}
+
+func (self *_Assembler) skip_one() {
+    self.Link(_LB_skip_one)                     // _skip:
+    self.Emit("MOVQ", _VAR_ic, _IC)             // MOVQ    _VAR_ic, IC
+    self.call_sf(_F_skip_one)                   // CALL_SF skip_one
+    self.Emit("TESTQ", _AX, _AX)                // TESTQ   AX, AX
+    self.Sjmp("JS"   , _LB_parsing_error_v)     // JS      _parse_error_v
+    self.Emit("MOVQ" , _VAR_pc, _R9)            // MOVQ    pc, R9
+    self.Rjmp("JMP"  , _R9)                     // JMP     (R9)
 }
 
 func (self *_Assembler) field_error() {
@@ -575,10 +636,23 @@ var (
     _F_vunsigned = jit.Imm(int64(native.S_vunsigned))
 )
 
-func (self *_Assembler) check_err() {
+func (self *_Assembler) check_err(vt reflect.Type) {
     self.Emit("MOVQ" , _VAR_st_Vt, _AX)         // MOVQ st.Vt, AX
     self.Emit("TESTQ", _AX, _AX)                // CMPQ AX, ${native.V_STRING}
-    self.Sjmp("JS"   , _LB_parsing_error_v)     // JNE  _parsing_error_v
+    // try to skip the value
+    if vt != nil {
+        self.Sjmp("JNS" , "_check_err_{n}")        // JNE  _parsing_error_v
+        self.Emit("MOVQ", _BP, _VAR_ic)
+        self.Emit("MOVQ", jit.Type(vt), _AX)  
+        self.Emit("MOVQ", _AX, _VAR_et)
+        self.Byte(0x4c  , 0x8d, 0x0d)         // LEAQ (PC), R9
+        self.Sref("_check_err_{n}", 4)
+        self.Emit("MOVQ", _R9, _VAR_pc)
+        self.Sjmp("JMP" , _LB_skip_one)
+        self.Link("_check_err_{n}")
+    } else {
+        self.Sjmp("JS"   , _LB_parsing_error_v)     // JNE  _parsing_error_v
+    }
 }
 
 func (self *_Assembler) check_eof(d int64) {
@@ -595,22 +669,25 @@ func (self *_Assembler) check_eof(d int64) {
 func (self *_Assembler) parse_string() {    // parse_string has a validate flag params in the last
     self.Emit("MOVQ", _ARG_fv, _CX)
     self.call_vf(_F_vstring)
-    self.check_err()
+    self.check_err(nil)
 }
 
 func (self *_Assembler) parse_number() {
+    self.Emit("MOVQ", _IC, _BP)
     self.call_vf(_F_vnumber)                               // call  vnumber
-    self.check_err()
+    self.check_err(floatType)
 }
 
 func (self *_Assembler) parse_signed() {
+    self.Emit("MOVQ", _IC, _BP)
     self.call_vf(_F_vsigned)
-    self.check_err()
+    self.check_err(intType)
 }
 
 func (self *_Assembler) parse_unsigned() {
+    self.Emit("MOVQ", _IC, _BP)
     self.call_vf(_F_vunsigned)
-    self.check_err()
+    self.check_err(uintType)
 }
 
 // Pointer: DI, Size: SI, Return: R9  
@@ -1144,7 +1221,17 @@ func (self *_Assembler) _asm_OP_bool(_ *_Instr) {
     self.Sjmp("JE"  , "_false_{n}")                             // JE   _false_{n}
     self.Emit("MOVL", jit.Imm(_IM_true), _CX)                   // MOVL $"true", CX
     self.Emit("CMPL", _CX, jit.Sib(_IP, _IC, 1, 0))             // CMPL CX, (IP)(IC)
-    self.Sjmp("JNE" , _LB_im_error)                             // JNE  _im_error
+    self.Sjmp("JE" , "_bool_true_{n}")          
+    // try to skip the value
+    self.Emit("MOVQ", _IC, _VAR_ic)           
+    self.Emit("MOVQ", _T_bool, _AX)  
+    self.Emit("MOVQ", _AX, _VAR_et)
+    self.Byte(0x4c, 0x8d, 0x0d)         // LEAQ (PC), R9
+    self.Sref("_end_{n}", 4)
+    self.Emit("MOVQ", _R9, _VAR_pc)
+    self.Sjmp("JMP"  , _LB_skip_one) 
+
+    self.Link("_bool_true_{n}")
     self.Emit("MOVQ", _AX, _IC)                                 // MOVQ AX, IC
     self.Emit("MOVB", jit.Imm(1), jit.Ptr(_VP, 0))              // MOVB $1, (VP)
     self.Sjmp("JMP" , "_end_{n}")                               // JMP  _end_{n}
@@ -1634,6 +1721,42 @@ func (self *_Assembler) _asm_OP_check_char(p *_Instr) {
     self.Emit("CMPB"   , jit.Sib(_IP, _IC, 1, 0), jit.Imm(int64(p.vb())))   // CMPB    (IP)(IC), ${p.vb()}
     self.Emit("CMOVQEQ", _AX, _IC)                                          // CMOVQEQ AX, IC
     self.Xjmp("JE"     , p.vi())                                            // JE      {p.vi()}
+}
+
+func (self *_Assembler) _asm_OP_check_char_0(p *_Instr) {
+    self.check_eof(1)
+    self.Emit("CMPB", jit.Sib(_IP, _IC, 1, 0), jit.Imm(int64(p.vb())))   // CMPB    (IP)(IC), ${p.vb()}
+    self.Xjmp("JE"  , p.vi())                                            // JE      {p.vi()}
+}
+
+func (self *_Assembler) _asm_OP_check_bool(p *_Instr) {
+    self.check_eof(1)
+    self.Emit("MOVBLZX", jit.Sib(_IP, _IC, 1, 0), _AX)       
+    self.Emit("CMPB", _AX, jit.Imm(int64('f')))  
+    self.Xjmp("JE"  , p.vi())               
+    self.Emit("CMPB", _AX, jit.Imm(int64('t')))
+    self.Xjmp("JE"  , p.vi())                               
+}
+
+func (self *_Assembler) _asm_OP_check_num(p *_Instr) {
+    self.check_eof(1)
+    self.Emit("MOVBLZX", jit.Sib(_IP, _IC, 1, 0), _AX)       
+    b := p.vb()
+    if b == 2 { // json.Number
+        self.Emit("CMPB", _AX, jit.Imm(int64('"')))  
+        self.Xjmp("JE"  , p.vi())     
+    }
+    if b == 1 { // negative number
+        self.Emit("CMPB", _AX, jit.Imm(int64('-')))  
+        self.Xjmp("JE"  , p.vi())     
+    } 
+    self.Emit("LEAL", jit.Ptr(_AX, -int64('0')), _CX)
+    self.Emit("CMPB", _CX, jit.Imm(int64('9'-'0')))  
+    self.Xjmp("JLS" , p.vi())   
+}
+
+func (self *_Assembler) _asm_OP_add(p *_Instr) {
+    self.Emit("ADDQ", jit.Imm(int64(p.vi())), _IC)  // ADDQ ${p.vi()}, IC
 }
 
 func (self *_Assembler) _asm_OP_load(_ *_Instr) {